Update browse-url-kde variables
* lisp/browse-url.el (browse-url-kde-program)
(browse-url-kde-args):
Update defaults to more modern alternative to kfmclient.
Kfmclient is only installed when Konqueror is installed which isn't the
default file-manager since KDE4.
In contrast it is more likely that kde-open is installed as it is
required by Plasma, it exists since KDE4.
`browse-url-kde-args` isn't required anymore but kept to not break
compatibility.