Refactor gnus-refer-thread-use-search
* lisp/gnus/gnus-sum.el (gnus-refer-thread-use-search):
Make it easier to customize the variable by adding predefine choices
as well as descriptions to each choice. Add option to add the value
`current' to the list of servers and groups to add the current group.
E.g. to add another group and the current group.
* lisp/gnus/gnus-search.el
(gnus-refer-thread-maybe-add-current-group, gnus-search-thread):
Add helper function to determine if the option contains
the value current. Use helper function.
* lisp/gnus/nnselect.el (nnselect-request-thread): Use helper
function to check if the value current is in
gnus-refer-thread-use-search.
* doc/misc/gnus.texi (gnus-refer-thread-use-search):
Include the current value in the section. Format the section
a little better to have a paragraph for each possible value that
the option may contain. Explain the option to have
a list of servers more detailed , i.e. similarly to the custom
description. Highlight that the current group is not searched
unless specified. Explain that unless the articles of that
thread are contained in the searched groups not adding
the current group might not be the desired effect.